Split the protocol module by concern
One 1660-line file had lost its organizing principle -- projectDataset had
come to sit between the SafeEvent property arrays and the SafeEvent target
arrays, which is accretion rather than any principle -- and a record's gate
sat hundreds of lines from the record it gates.

The dependency graph was already an acyclic layering, so the cut needed no
untangling: cloneable and refusal depend on nothing; untrusted-input reads a
producer's record safely; version gates the envelope; each record file owns
its shape and, where it has one, its gate; runtime names the operations;
child-formats is independent.

boxel-execution-protocol.ts stays as the barrel. runtime-common's
`"./*": "./*.ts"` export maps a file rather than a directory, so keeping it
preserves the public import path -- and unlike runtime-common's own index.ts
this barrel is harmless, since every file behind it holds the same
properties.

Pure motion: no behavior change, and the test count is the check -- 42
tests, 188 assertions before and after. Helpers that now cross a file
boundary are exported; nothing else changed.

The import-closure guard records the new file set, and caught its own
omission when version.ts was missing from it.
