+++ packages/runtime-common/boxel-execution-protocol/instance-projection.ts
@@ -1,0 +1,207 @@
+/**
+ * One instance's state as data, with linked values as references rather than
+ * expanded graphs.
+ *
+ * Part of the cross-boundary execution protocol (RP-14); see
+ * `../boxel-execution-protocol.ts` for the properties every file here holds.
+ */
+
+import type * as JSONTypes from 'json-typescript';
+
+import { isCodeRef } from '../card-document-shape.ts';
+import type { CodeRef } from '../code-ref.ts';
+import type { RealmResourceIdentifier } from '../realm-identifiers.ts';
+import type { Cloneable } from './cloneable.ts';
+import { isPlainRecord, readMember } from './untrusted-input.ts';
+import type { ProtocolEnvelope } from './version.ts';
+
+/**
+ * How a linked value appears in a projection: an identity and a type, which
+ * the recipient resolves through the canonical Store — never the linked
+ * card's own data. A card holding a link to a card holding a link does not
+ * hand its recipient a graph to walk.
+ */
+export type BoxelValueReference = Cloneable<{
+  $boxel: {
+    id: RealmResourceIdentifier | null;
+    type: CodeRef;
+  };
+}>;
+
+/**
+ * Whether a projected value is a link reference rather than data.
+ *
+ * Exact by design: a `$boxel` marker carrying anything beyond `id` and
+ * `type`, or sitting beside sibling members, is an expanded graph wearing a
+ * reference's clothes and answers `false` here rather than being accepted as
+ * a reference.
+ *
+ * `type` is checked as a real `CodeRef`, not merely as an object, because
+ * this predicate narrows and its caller resolves that ref through the Store.
+ * A structurally plausible ref — `{}`, a `{module}` with no `name`, an
+ * `ancestorOf` whose `card` is garbage — would otherwise arrive there
+ * carrying the authority of a check that never happened.
+ */
+export function isBoxelValueReference(
+  value: unknown,
+): value is BoxelValueReference {
+  try {
+    if (!isPlainRecord(value) || !hasExactOwnKeys(value, ['$boxel'])) {
+      return false;
+    }
+    let marker = readMember(value, '$boxel');
+    if (!isPlainRecord(marker) || !hasExactOwnKeys(marker, ['id', 'type'])) {
+      return false;
+    }
+    let id = readMember(marker, 'id');
+    return (
+      (id === null || typeof id === 'string') &&
+      isExactCodeRef(readMember(marker, 'type'))
+    );
+  } catch {
+    // A predicate whose contract is to answer must not throw instead: a
+    // marker built from a throwing accessor is simply not a reference.
+    return false;
+  }
+}
+
+/**
+ * Whether an object's own property names are exactly `expected` — enumerable
+ * or not.
+ *
+ * `Object.keys` would skip a non-enumerable member, and a member the check
+ * skipped is still reachable by whoever holds the object. That is how an
+ * entire card rides inside a value that answers "reference".
+ */
+function hasExactOwnKeys(source: object, expected: string[]): boolean {
+  let names = Object.getOwnPropertyNames(source);
+  if (names.length !== expected.length) {
+    return false;
+  }
+  let sorted = [...names].sort();
+  let wanted = [...expected].sort();
+  return (
+    sorted.every((name, index) => name === wanted[index]) &&
+    Object.getOwnPropertySymbols(source).length === 0
+  );
+}
+
+/**
+ * How deep an `ancestorOf` / `fieldOf` chain may nest before it is refused.
+ * Real refs are one or two levels; the bound exists because the value is the
+ * far side's to shape.
+ */
+const MAX_CODE_REF_DEPTH = 16;
+
+/**
+ * A code ref carrying its own members and nothing else.
+ *
+ * Stricter than `isCodeRef`, deliberately. That predicate answers "can this be
+ * read as a ref", which is right for a document whose resources may carry
+ * more than one reader needs; here the question is whether a value is a
+ * reference *instead of* data. A ref admitting extra members lets an entire
+ * card ride inside `type` — the expanded graph the marker's own exactness
+ * check was written to refuse, one level further down.
+ */
+function isExactCodeRef(ref: unknown, depth = MAX_CODE_REF_DEPTH): boolean {
+  // A predicate whose contract is to answer must not throw instead. The
+  // traversal is this function's own, and `isCodeRef` is only ever handed a
+  // leaf — it recurses without a bound of its own, so a nested chain would
+  // blow the stack inside it before this guard could apply.
+  if (depth <= 0 || !isPlainRecord(ref)) {
+    return false;
+  }
+  // A ref inheriting a discriminator reads as one form here and another at
+  // the Store, so the prototype is part of the shape.
+  let prototype = Object.getPrototypeOf(ref);
+  if (prototype !== Object.prototype && prototype !== null) {
+    return false;
+  }
+  let type = readMember(ref, 'type');
+  if (type === 'ancestorOf') {
+    return (
+      hasExactOwnKeys(ref, ['card', 'type']) &&
+      isExactCodeRef(readMember(ref, 'card'), depth - 1)
+    );
+  }
+  if (type === 'fieldOf') {
+    return (
+      hasExactOwnKeys(ref, ['card', 'field', 'type']) &&
+      typeof readMember(ref, 'field') === 'string' &&
+      isExactCodeRef(readMember(ref, 'card'), depth - 1)
+    );
+  }
+  // The leaf's own member types are the repo's predicate to judge, not a
+  // second opinion written here.
+  return hasExactOwnKeys(ref, ['module', 'name']) && isCodeRef(ref);
+}
+
+/**
+ * One instance's state, as data.
+ *
+ * `model` carries declared field values and the JSON-safe results of getters
+ * the owning runtime evaluated. Linked values appear as
+ * `BoxelValueReference`s — the projection is one instance deep, never an
+ * expanded graph. The type system cannot state that rule (a reference is
+ * structurally an object like any other), so the producer holds it: the
+ * projection pipeline knows each field's kind and emits a reference for every
+ * `linksTo`/`linksToMany`. `isBoxelValueReference` is how a consumer tells
+ * the two apart.
+ *
+ * `revision` orders projections of one instance against each other. It is not
+ * an etag and carries no server meaning — main's write path has no revision
+ * token (RP-9.6) — it exists so a recipient can drop a projection that a
+ * newer one has already superseded in flight.
+ */
+export type InstanceProjection = Cloneable<
+  ProtocolEnvelope & {
+    id: RealmResourceIdentifier | null;
+    type: CodeRef;
+    revision: number;
+    model: Record<string, JSONTypes.Value>;
+    presentation: InstancePresentation;
+  }
+>;
+
+/**
+ * What the Host's own chrome needs in order to wrap an instance, derived
+ * Host-side and crossed as data.
+ *
+ * The theme members are why this record exists rather than being read out of
+ * `model`. A themed card's stylesheet lives on a *linked* Theme card, which a
+ * projection carries only as a reference — resolving it is exactly the graph
+ * walk the projection forbids. So the Host resolves the theme once, against
+ * the canonical instance, and the three derived strings cross:
+ *
+ * - `themeScope` is the `data-boxel-theme-scope` token, a content hash of the
+ *   theme's id and CSS (RP-11.3), so shared themes emit one stylesheet and
+ *   prerendered HTML stays stable across processes.
+ * - `themeCss` is the theme's raw custom-property block, from which the
+ *   scoped stylesheet compiles.
+ * - `cssImports` are the stylesheet imports the theme depends on, typically
+ *   font faces.
+ *
+ * A tier makes the same trusted `CardContainer` invocation main makes from
+ * these; without them a themed card renders unthemed, which is not a degraded
+ * theme but a different design.
+ *
+ * `isThemed` is carried rather than derived, because neither `theme` nor
+ * `themeCss` implies it. Base answers the question two different ways
+ * (`hasTheme`, `field-component.gts`): an ordinary card is themed when it
+ * links a Theme, but a Theme card previewing its own CSS is themed when that
+ * CSS is non-empty — and such a card links no Theme at all, so `theme` is
+ * `null` while the three derived strings are not. Reading `theme !== null` as
+ * "themed" renders a Theme card's own preview without the CSS it exists to
+ * show; reading `themeCss !== null` gets the converse wrong, since a card
+ * linking a Theme whose variables are empty is still themed.
+ */
+export type InstancePresentation = Cloneable<{
+  title: string | null;
+  summary: string | null;
+  thumbnailURL: string | null;
+  isThemed: boolean;
+  theme: BoxelValueReference | null;
+  themeScope: string | null;
+  themeCss: string | null;
+  cssImports: string[] | null;
+}>;
